requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
tgcaller==1.0.3
yt-dlp==2024.10.7
python-dotenv==1.0.1
loguru==0.7.2
